Had some progress in the backend

+++ PasswordManager/src/main/java/com/example/PasswordManager/entity/Password.java
@@ -7,7 +7,7 @@ public class Password {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String siteName;
     private String URL;
 
@@ -21,11 +21,11 @@ public class Password {
 
 
     // Getters and Setters
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
