+++ PasswordManager/src/main/java/com/example/PasswordManager/repository/PasswordRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PasswordRepo extends JpaRepository<Password, Long> {
+public interface PasswordRepo extends JpaRepository<Password, Integer> {
     List<Password> findAllByUserEmail(String email);
 }
